third_party/WebKit/LayoutTests/presentation/presentation-request-iframe-default-success.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<iframe></iframe>
<script>

var iframe = document.querySelector('iframe');

async_test(t => {
  iframe.onload = t.step_func(() => {
    iframe.contentWindow.postMessage('create', '*');
  });

  window.addEventListener('message', t.step_func_done(e => {
    assert_equals(e.data, 'success');
  }));

  // Navigate the iframe
  iframe.src = 'resources/iframe-sandbox.html';
}, "Test that creating PresentationRequest succeeds when the iframe is not sandboxed.");

</script>
</body>
</html>
